bot_telegram/requirements.txt
# Bot de Telegram
python-telegram-bot>=20.7
python-dotenv>=1.0.0

# Deep Learning y Visión
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.8.1.78
numpy
pillow>=10.1.0

# Detección de Postura
mediapipe
# mmpose==1.3.0
# mmcv==2.1.0
# mmengine==0.10.1

# Procesamiento de Video
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3

# Utilidades
requests
aiohttp>=3.9.1
matplotlib>=3.8.2

# Servidor HTTP
fastapi
uvicorn
python-multipart
